Added linux run_server, TV Series sorting.
Updating styling to make room for a comment section, added TV series
sorting which uses regular expressions to determine episode and season
number, which is used to sort the playlist. Videos are now taken from
the channel's uploads, rather than a search with a channel filter, which
dropped some videos. An input with the current playlist index has been
added, and theoretically the last video played should be started
automatically on reopening the page.